Refactor inventory
