+++ finish/src/main/java/io/openliberty/guides/inventory/InventoryManager.java
@@ -17,6 +17,10 @@ import java.net.URL;
 import java.net.UnknownHostException;
 import java.net.MalformedURLException;
 import javax.ws.rs.ProcessingException;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import java.util.Properties;
 import javax.inject.Inject;
 import javax.enterprise.context.ApplicationScoped;
@@ -24,6 +28,7 @@ import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.eclipse.microprofile.rest.client.RestClientBuilder;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 import io.openliberty.guides.inventory.model.InventoryList;
+import io.openliberty.guides.inventory.model.SystemData;
 import io.openliberty.guides.inventory.client.SystemClient;
 import io.openliberty.guides.inventory.client.UnknownUrlException;
 import io.openliberty.guides.inventory.client.UnknownUrlExceptionMapper;
@@ -31,7 +36,7 @@ import io.openliberty.guides.inventory.client.UnknownUrlExceptionMapper;
 @ApplicationScoped
 public class InventoryManager {
 
-  private InventoryList invList = new InventoryList();
+  private List<SystemData> systems = Collections.synchronizedList(new ArrayList<>());
   private final String DEFAULT_PORT = System.getProperty("default.http.port");
 
   @Inject
@@ -39,7 +44,6 @@ public class InventoryManager {
   private SystemClient defaultRestClient;
 
   public Properties get(String hostname) {
-
     Properties properties = null;
     if (hostname.equals("localhost")) {
       properties = getPropertiesWithDefaultHostName();
@@ -47,14 +51,21 @@ public class InventoryManager {
       properties = getPropertiesWithGivenHostName(hostname);
     }
 
-    if (properties != null) {
-      invList.addToInventoryList(hostname, properties);
-    }
     return properties;
   }
 
+  public void add(String hostname, Properties systemProps) {
+    Properties props = new Properties();
+    props.setProperty("os.name", systemProps.getProperty("os.name"));
+    props.setProperty("user.name", systemProps.getProperty("user.name"));
+
+    SystemData host = new SystemData(hostname, props);
+    if (!systems.contains(host))
+      systems.add(host);
+  }
+
   public InventoryList list() {
-    return invList;
+    return new InventoryList(systems);
   }
 
   private Properties getPropertiesWithDefaultHostName() {
