+++ finish/src/main/java/io/openliberty/guides/inventory/model/SystemData.java
@@ -1,0 +1,44 @@
+// tag::copyright[]
+/*******************************************************************************
+ * Copyright (c) 2017, 2018 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     IBM Corporation - Initial implementation
+ *******************************************************************************/
+// end::copyright[]
+package io.openliberty.guides.inventory.model;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Properties;
+
+public class SystemData {
+
+    private final String hostname;
+    private final Properties properties;
+
+    public SystemData(String hostname, Properties properties) {
+        this.hostname = hostname;
+        this.properties = properties;
+    }
+
+    public String getHostname() {
+        return hostname;
+    }
+
+    public Properties getProperties() {
+        return properties;
+    }
+
+    @Override
+    public boolean equals(Object host) {
+        if (host instanceof SystemData) {
+        return hostname.equals(((SystemData) host).getHostname());
+        }
+        return false;
+    }
+}
